+++ packages/dd-trace/src/proxy.js
@@ -46,7 +46,7 @@ class Tracer extends BaseTracer {
 
           // dirty require for now so zero appsec code is executed unless explicitely enabled
           if (config.appsec && config.appsec.enabled) {
-            require('../../dd-appsec').enable(config)
+            require('./appsec').enable(config)
           }
 
           this._tracer = new DatadogTracer(config)
